fix(mxc): make the demo example unix-only; build windows in CI
Two halves of the same gap.

examples/mxc_demo.rs is a demo of an AF_UNIX protocol, so it cannot
compile on windows -- and Cargo cannot target-gate an [[example]]. The
body moves to examples/mxc_demo_support/imp.rs, included via #[path]
behind cfg(unix), with a stub main elsewhere. A subdirectory without a
main.rs is not auto-discovered, so dump_theme/libcheck/radiocheck keep
being found -- setting autoexamples = false would have silently
de-registered all three.

CI gains windows-latest. The matrix was ubuntu + macos only, which is
precisely why an ungated UnixStream got through review and a green run.

Verified: windows lib check passes, linux 55/55, --no-default-features
compiles, default bin builds, fmt + clippy clean.
