+++ .github/workflows/ci.yml
@@ -37,7 +37,12 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        # Windows is a supported target (Cargo.toml has a cfg(windows) section
+        # and souvlaki ships a Windows backend), but it was never built in CI —
+        # which is exactly how an ungated `std::os::unix::net::UnixStream` in
+        # the MXC layer reached a PR. Building it here makes that class of
+        # regression impossible to merge.
+        os: [ubuntu-latest, macos-latest, windows-latest]
     steps:
       - uses: actions/checkout@v4
